+++ tests/data/fake_large_chromosome/tracks.conf
@@ -4,5 +4,11 @@ csiUrlTemplate=test.vcf.gz.csi
 storeTimeout=100
 storeClass=JBrowse/Store/SeqFeature/VCFTabix
 type=CanvasVariants
+[tracks.bam]
+urlTemplate=out.bam
+csiUrlTemplate=out.bam.csi
+storeTimeout=100
+storeClass=JBrowse/Store/SeqFeature/BAM
+type=Alignments2
 [GENERAL]
 classicMenu=true
